% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper1.R
\name{get_numeric_variables}
\alias{get_numeric_variables}
\title{Return the numeric categorical variables}
\usage{
get_numeric_variables(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Return the numeric categorical variables
}
