% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histopathRshiny-package.R
\docType{package}
\name{histopathRshiny-package}
\alias{histopathRshiny}
\alias{histopathRshiny-package}
\title{histopathRshiny: Histopathology Research Template Shiny App}
\description{
Histopathology Research Template Shiny App.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sbalci/histopathRshiny}
  \item Report bugs at \url{https://github.com/sbalci/histopathRshiny/issues}
}

}
\author{
\strong{Maintainer}: Serdar Balci \email{drserdarbalci@gmail.com} (\href{https://orcid.org/0000-0002-7852-3851}{ORCID})

}
\keyword{internal}
