% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_interval.R
\name{time_interval}
\alias{time_interval}
\title{Title}
\usage{
time_interval(mydata, SurgeryDate, LastFollowUpDate)
}
\arguments{
\item{mydata}{}

\item{SurgeryDate}{}

\item{LastFollowUpDate}{}
}
\value{

}
\description{
Title
}
